refactor: rename utils file to misc
