+++ src/django_components/component.py
@@ -62,7 +62,7 @@ from django_components.slots import (
 )
 from django_components.template import cached_template
 from django_components.util.logger import trace_msg
-from django_components.util.utils import gen_id
+from django_components.util.misc import gen_id
 from django_components.util.validation import validate_typed_dict, validate_typed_tuple
 
 # TODO_REMOVE_IN_V1 - Users should use top-level import instead
