+++ src/django_components/context.py
@@ -10,7 +10,7 @@ from typing import Any, Dict, Optional
 
 from django.template import Context, TemplateSyntaxError
 
-from django_components.util.utils import find_last_index
+from django_components.util.misc import find_last_index
 
 _FILLED_SLOTS_CONTENT_CONTEXT_KEY = "_DJANGO_COMPONENTS_FILLED_SLOTS"
 _ROOT_CTX_CONTEXT_KEY = "_DJANGO_COMPONENTS_ROOT_CTX"
