+++ src/django_components/provide.py
@@ -8,7 +8,7 @@ from django_components.context import set_provided_context_var
 from django_components.expression import RuntimeKwargs
 from django_components.node import BaseNode
 from django_components.util.logger import trace_msg
-from django_components.util.utils import gen_id
+from django_components.util.misc import gen_id
 
 PROVIDE_NAME_KWARG = "name"
